@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 3.0.0+10
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <=3.3.0"
 
 dependencies:
   cloud_firestore: ^4.3.2
@@ -46,16 +46,14 @@ dependencies:
   firebase_messaging: ^14.2.5
   flutter_local_notifications: ^13.0.0
   geolocator: ^9.0.2
-<<<<<<< HEAD
   flutter_config: ^2.0.0
-=======
   flame: 1.6.0
   flame_audio: ^1.4.0
   flutter_carousel_slider: 1.1.0
   hive: 2.2.3
   hive_generator: 2.0.0
   path_provider: 2.0.11
->>>>>>> d803f81 ([ADD] Game-Spacescapes)
+  
   flutter:
     sdk: flutter
 
