@@ -2,19 +2,19 @@
 /// to create an [Enemy] component.
 class EnemyData {
   // Speed of the enemy.
-  final double speed;
+  final double? speed;
 
   // Sprite ID from the main sprite sheet.
-  final int spriteId;
+  final int? spriteId;
 
   // Level of this enemy.
-  final int level;
+  final int? level;
 
   // Indicates if this enemy can move horizontally.
-  final bool hMove;
+  final bool? hMove;
 
   // Points gains after destroying this enemy.
-  final int killPoint;
+  final int? killPoint;
 
   const EnemyData({
      this.speed,
