@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 class SABT extends StatefulWidget {
   final Widget child;
   const SABT({
-    Key key,
-    @required this.child,
+    Key? key,
+    required this.child,
   }) : super(key: key);
   @override
   _SABTState createState() {
@@ -11,8 +11,8 @@ class SABT extends StatefulWidget {
   }
 }
 class _SABTState extends State<SABT> {
-  ScrollPosition _position;
-  bool _visible;
+  ScrollPosition? _position;
+  bool? _visible;
   @override
   void dispose() {
     _removeListener();
@@ -33,7 +33,7 @@ class _SABTState extends State<SABT> {
     _position?.removeListener(_positionListener);
   }
   void _positionListener() {
-    final FlexibleSpaceBarSettings settings =
+    final FlexibleSpaceBarSettings? settings =
     context.dependOnInheritedWidgetOfExactType<FlexibleSpaceBarSettings>();
     bool visible = settings == null || settings.currentExtent <= settings.minExtent;
     if (_visible != visible) {
@@ -45,7 +45,7 @@ class _SABTState extends State<SABT> {
   @override
   Widget build(BuildContext context) {
     return Visibility(
-      visible: _visible,
+      visible: _visible!,
       child: widget.child,
     );
   }
