@@ -6,22 +6,24 @@ class CustomGestureDetector extends StatelessWidget{
   static const int AXIS_Y = 1;
   static const int AXIS_BOTH = 2;
 
-  final int axis;
+  final int? axis;
   final Widget child;
   final double velocity;
-  final Function onSwipeUp;
-  final Function onSwipeDown;
-  final Function onSwipeLeft;
-  final Function onSwipeRight;
+  final Function? onSwipeUp;
+  final Function? onSwipeDown;
+  final Function? onSwipeLeft;
+  final Function? onSwipeRight;
   final Function onTap;
 
-  CustomGestureDetector({@required this.onTap,@required this.child,@required this.velocity,this.onSwipeLeft,this.onSwipeRight,this.onSwipeUp,this.onSwipeDown,@required this.axis});
+  CustomGestureDetector({required this.onTap,required this.child,required this.velocity,this.onSwipeLeft,this.onSwipeRight,this.onSwipeUp,this.onSwipeDown,@required this.axis});
 
   @override
   Widget build(BuildContext context) {
     // TODO: implement build
     return GestureDetector(
-        onTap: onTap,
+        onTap: (){
+          onTap();
+        },
         onPanEnd: (details){
 
           Offset v = details.velocity.pixelsPerSecond;
@@ -29,17 +31,17 @@ class CustomGestureDetector extends StatelessWidget{
           try {
             if((axis == AXIS_Y) || axis == AXIS_BOTH){
               if(v.dy>velocity){
-                onSwipeDown();
+                onSwipeDown!();
               } else if(v.dy<-velocity){
-                onSwipeUp();
+                onSwipeUp!();
               }
             }
 
             if((axis == AXIS_X )  || axis == AXIS_BOTH){
               if(v.dx>velocity){
-                onSwipeRight();
+                onSwipeRight!();
               } else if(v.dx<-velocity) {
-                onSwipeLeft();
+                onSwipeLeft!();
               }
             }
           } catch (e){
